chip-fel-flash: put uboot environment into a UBI volume
You can pass 2 different uboot images. One is used for the production, and
should be built with ENV_IS_NOWHERE to speedup the flashing procedure.
The other one should be build with ENV_IS_IN_UBI.

You can also use the same image, but it will take longer to flash an
image because UBI metadata are written twice.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>